pyproject.toml
[tool.poetry]
name = "astutus"
version = "0.1.0"
description = ""
authors = ["lxmcneill <lxmcneill@gmail.com>"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/lxmcneill/astutus"
keywords = ["discord", "bot"]

[tool.poetry.dependencies]
python = "^3.8"
arrow = "^0.13.2"
"discord.py" = "^1.1"

[tool.poetry.dev-dependencies]
black = "^19.3b0"
typed-ast = {path = ".\\whl\\typed_ast-1.3.5-cp38-cp38m-win_amd64.whl"}
pylint = "^2.3"

[tool.poetry.extras]
uvloop = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
run = "manage:main"